Backend/requirements.txt
aniso8601==9.0.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
checksumdir==1.2.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
cryptography==41.0.2
cycler==0.11.0
dacite==1.8.1
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dnspython==2.2.1
email-validator==1.2.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
Flask==2.2.2
Flask-JWT==0.3.2
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fonttools==4.33.3
fqdn==1.5.1
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.1
htmlmin==0.1.12
idna==3.3
ImageHash==4.3.1
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
jwt==1.3.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==3.0.1
multimethod==1.9.1
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.22.4
overrides==7.4.0
packaging==21.3
pandas==1.4.2
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pdfminer.six==20221105
pdfplumber==0.10.1
phik==0.12.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.0.9
pypdf==3.12.2
pypdfium2==4.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-env==1.0.0
python-json-logger==2.0.7
pytz==2022.6
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
redis==5.0.1
referencing==0.30.2
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
scipy==1.11.2
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.45
stack-data==0.6.2
statsmodels==0.14.0
tangled-up-in-unicode==0.2.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.11
virtualenv==20.24.5
visions==0.7.5
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.2.2
widgetsnbextension==4.0.8
wordcloud==1.9.2
WTForms==3.0.1
XlsxWriter==3.1.2
ydata-profiling==4.5.1
